avcodec/lcevc_parser: Check that block_size is not negative
Based on 248b481c337d917e0fece9555ae59b5375c725c6
